+++ src/main/java/cl/praxis/bootcapp/controllers/rests/RoleRestController.java
@@ -1,4 +1,4 @@
-package cl.praxis.bootcapp.controllers;
+package cl.praxis.bootcapp.controllers.rests;
 
 import cl.praxis.bootcapp.entities.Role;
 import cl.praxis.bootcapp.services.imp.RoleServiceImp;
